raster: Fix multiplications converted to a larger type
Cast the first operand so that integer products used as allocation
sizes, file offsets, and pointer offsets are computed in size_t,
off_t, or the destination integer type instead of int, and so that
float products consumed as double (exp, log2, sin arguments and
double variables) are computed in double.

Fixes 36 CodeQL cpp/integer-multiplication-cast-to-long alerts in 15
raster tools.

+++ raster/r.distance/report.c
@@ -79,8 +79,9 @@ void report(struct Parms *parms)
     }
 
     if (parms->sort > 0)
-        lines = (struct ReportLine *)G_malloc(
-            map1->edges.ncats * map2->edges.ncats * sizeof(struct ReportLine));
+        lines = (struct ReportLine *)G_malloc((size_t)map1->edges.ncats *
+                                              map2->edges.ncats *
+                                              sizeof(struct ReportLine));
     else
         lines = NULL;
     nlines = 0;
